stock/sms/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
jose
passlib[bcrypt]
jinja2
python-jose